server/requirements.txt
astroid==3.0.2
bcrypt==4.1.2
bidict==0.22.1
blinker==1.7.0
click==8.1.7
dill==0.3.7
dnspython==2.4.2
Flask==3.0.0
Flask-Cors==4.0.0
Flask-SocketIO==5.3.6
gunicorn==21.2.0
h11==0.14.0
isort==5.13.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mccabe==0.7.0
packaging==23.2
Pillow==10.1.0
platformdirs==4.1.0
PyJWT==2.8.0
pylint==3.0.3
pymongo==4.6.1
python-dotenv==1.0.0
python-engineio==4.8.1
python-socketio==5.10.0
redis==5.0.1
simple-websocket==1.0.0
tomlkit==0.12.3
Werkzeug==3.0.1
wsproto==1.2.0
